+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/view/AbstractItemGridResourceType.java
@@ -1,10 +1,10 @@
 package com.refinedmods.refinedstorage.common.grid.view;
 
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
-import com.refinedmods.refinedstorage.api.resource.repository.ResourceRepositoryMapper;
 import com.refinedmods.refinedstorage.common.api.grid.GridResourceAttributeKeys;
 import com.refinedmods.refinedstorage.common.api.grid.view.GridResource;
 import com.refinedmods.refinedstorage.common.api.grid.view.GridResourceAttributeKey;
+import com.refinedmods.refinedstorage.common.api.grid.view.GridResourceType;
 import com.refinedmods.refinedstorage.common.support.resource.ItemResource;
 import com.refinedmods.refinedstorage.common.util.ClientPlatformUtil;
 
@@ -17,6 +17,7 @@ import java.util.stream.Stream;
 
 import com.google.common.base.Supplier;
 import com.google.common.base.Suppliers;
+import com.mojang.serialization.MapCodec;
 import net.minecraft.core.Holder;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.chat.Component;
@@ -27,8 +28,13 @@ import net.minecraft.world.item.TooltipFlag;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public abstract class AbstractItemGridResourceRepositoryMapper implements ResourceRepositoryMapper<GridResource> {
-    private static final Logger LOGGER = LoggerFactory.getLogger(AbstractItemGridResourceRepositoryMapper.class);
+public abstract class AbstractItemGridResourceType implements GridResourceType {
+    private static final Logger LOGGER = LoggerFactory.getLogger(AbstractItemGridResourceType.class);
+
+    @Override
+    public MapCodec<GridResource> getMapCodec() {
+        return GridResourceCodecs.ITEM;
+    }
 
     @Override
     public GridResource apply(final ResourceKey resource) {
@@ -52,6 +58,11 @@ public abstract class AbstractItemGridResourceRepositoryMapper implements Resour
             originalName,
             k -> attributes.getOrDefault(k, Collections::emptySet).get()
         );
+    }
+
+    @Override
+    public Class<? extends ResourceKey> getResourceKeyClass() {
+        return ItemResource.class;
     }
 
     private String getTooltip(final ItemStack itemStack) {