+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/view/FluidGridResource.java
@@ -11,6 +11,7 @@ import com.refinedmods.refinedstorage.common.api.grid.strategy.GridScrollingStra
 import com.refinedmods.refinedstorage.common.api.grid.view.AbstractGridResource;
 import com.refinedmods.refinedstorage.common.api.grid.view.GridResource;
 import com.refinedmods.refinedstorage.common.api.grid.view.GridResourceAttributeKey;
+import com.refinedmods.refinedstorage.common.api.grid.view.GridResourceType;
 import com.refinedmods.refinedstorage.common.api.support.resource.FluidOperationResult;
 import com.refinedmods.refinedstorage.common.api.support.resource.ResourceRendering;
 import com.refinedmods.refinedstorage.common.api.support.resource.ResourceType;
@@ -63,6 +64,11 @@ public class FluidGridResource extends AbstractGridResource<FluidResource> {
                 result.container(),
                 null
             )).stream().toList();
+    }
+
+    @Override
+    public GridResourceType getType() {
+        return Platform.INSTANCE.getFluidGridResourceType();
     }
 
     @Nullable
@@ -120,11 +126,6 @@ public class FluidGridResource extends AbstractGridResource<FluidResource> {
     }
 
     @Override
-    public boolean belongsToResourceType(final ResourceType resourceType) {
-        return resourceType == ResourceTypes.FLUID;
-    }
-
-    @Override
     public List<Component> getTooltip() {
         return Platform.INSTANCE.getFluidRenderer().getTooltip(resource);
     }
@@ -134,6 +135,14 @@ public class FluidGridResource extends AbstractGridResource<FluidResource> {
         return Optional.empty();
     }
 
+    public FluidResource getFluidResource() {
+        return fluidResource;
+    }
+
+    @Override
+    public boolean is(final ResourceType resourceType) {
+        return resourceType == ResourceTypes.FLUID;
+    }
 
     @Override
     public boolean is(final GridResource other) {